Ejercicios 9 y 9 bis, 10 y 10 bis resueltos entrada y salida

+++ 00-Unidades/01_entrada_salida/Ejercicios_entrada_salida/ES_app_09.py
@@ -33,7 +33,7 @@ class App(customtkinter.CTk):
 
 
     def btn_mostrar_on_click(self):
-        sueldo = int(self.txt_sueldo.get())
+        sueldo = float(self.txt_sueldo.get())
         aumento = "15 por ciento"
         sueldo_con_aumento = round(sueldo * 1.15)
         alert("Sueldo", "Tu sueldo actualizado es: {}" .format(sueldo_con_aumento))
